WC-5233 restructure commands
