+++ .changeset/preview-secret-per-preview.md
@@ -2,6 +2,6 @@
 "wrangler": minor
 ---
 
-Add `wrangler preview secret put`, `delete`, `list`, and `bulk` for managing secrets on a named Worker Preview
+Use Preview deployment PATCH APIs for Preview secret commands
 
-These private-beta commands manage secrets on a single named Preview's latest deployment, without affecting production, other Previews, or the Worker's Previews base configuration. `put`, `delete`, and `bulk` create a new Preview deployment that goes live at 100% immediately and report the live Preview URL on success; `list` reads the latest deployment and prints secret names with values masked. Each command accepts `--name` to target a Preview (defaulting to the current git branch).
+Wrangler now updates Worker Preview secrets by patching the named Preview's latest deployment instead of patching the Worker's Previews settings. This keeps secret changes scoped to one Preview, avoids affecting production or other Previews, and creates a new Preview deployment that goes live at 100% immediately. `list` now reads from the named Preview's latest deployment and prints secret names with values masked.
